requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==3.7.1
attrs==25.4.0
Authlib==1.6.6
bcrypt==5.0.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
dataclasses-json==0.6.7
ecdsa==0.19.1
fastapi==0.104.1
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.0.340
langchain-core==0.1.23
langgraph==0.0.15
langsmith==0.0.87
MarkupSafe==3.0.3
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
networkx==3.2.1
numpy==1.24.3
packaging==23.2
passlib==1.7.4
propcache==0.4.1
pyasn1==0.6.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.0.0
python-jose==3.5.0
PyYAML==6.0.3
regex==2025.11.3
requests==2.31.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.11.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.27.0
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.15.2
torch==2.1.1
tqdm==4.67.1
transformers==4.35.2
tree-sitter-python==0.23.0
tree_sitter==0.20.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.24.0
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
